# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales Manager API",
        "description": "This API manages sales operations, including consulting, creating, updating, and deleting clients and products, as well as querying sales data.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://0.0.0.0:8000"
        }
    ],
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "List all clients",
                "operationId": "d00d20521584ab5da83e088abb7aded2",
                "responses": {
                    "200": {
                        "description": "List of clients retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create a new client",
                "operationId": "f1e86e5b79fedfa6593067977292b13b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf",
                                    "email",
                                    "phone",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "janedoe@email.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "12991234567"
                                    },
                                    "address": {
                                        "properties": {
                                            "street": {
                                                "type": "string",
                                                "example": "Rua das Flores"
                                            },
                                            "number": {
                                                "type": "string",
                                                "example": "123"
                                            },
                                            "complement": {
                                                "type": "string",
                                                "example": "Apto 101"
                                            },
                                            "neighbourhood": {
                                                "type": "string",
                                                "example": "Centro"
                                            },
                                            "city": {
                                                "type": "string",
                                                "example": "São Paulo"
                                            },
                                            "state": {
                                                "type": "string",
                                                "example": "SP"
                                            },
                                            "postal_code": {
                                                "type": "string",
                                                "example": "01000000"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Client created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Client"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients/{id}/sales": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get all client's sales by date (descending)",
                "operationId": "48dac8f79272d4465be7caf20b11cdd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client sales retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Sale"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients/{id}/sales/{year}/{month}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get a client's sales by month and year",
                "operationId": "69dde9dccce110f53ac53db11689e007",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "path",
                        "description": "Year",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "month",
                        "in": "path",
                        "description": "Month",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client sales retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Sale"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid month or year provided."
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "title": "Client",
                "description": "Client model",
                "properties": {
                    "id": {
                        "description": "Client ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Client name",
                        "type": "string"
                    },
                    "cpf": {
                        "description": "Client CPF",
                        "type": "string"
                    },
                    "email": {
                        "description": "Client email",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Client phone",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "address": {
                        "properties": {
                            "street": {
                                "description": "Street name",
                                "type": "string"
                            },
                            "number": {
                                "description": "Street number",
                                "type": "string"
                            },
                            "complement": {
                                "description": "Address complement",
                                "type": "string"
                            },
                            "neighbourhood": {
                                "description": "Neighbourhood",
                                "type": "string"
                            },
                            "city": {
                                "description": "City",
                                "type": "string"
                            },
                            "state": {
                                "description": "State",
                                "type": "string"
                            },
                            "postal_code": {
                                "description": "Postal code",
                                "type": "string"
                            },
                            "created_at": {
                                "description": "Creation timestamp",
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "description": "Update timestamp",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Sale": {
                "title": "Sale",
                "description": "Sale model",
                "properties": {
                    "id": {
                        "description": "Sale ID",
                        "type": "integer"
                    },
                    "client_id": {
                        "description": "Client ID",
                        "type": "integer"
                    },
                    "product_id": {
                        "description": "Product ID",
                        "type": "integer"
                    },
                    "quantity": {
                        "description": "Quantity of products sold",
                        "type": "integer"
                    },
                    "unit_price": {
                        "description": "Unit price of the product",
                        "type": "number",
                        "format": "float"
                    },
                    "total_price": {
                        "description": "Total price of the sale",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}